feat: доработка графов бизнес процесса FRO-861
